backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
python-multipart==0.0.6
boto3==1.34.0
reportlab==4.0.8
pypdf==3.17.0
pydantic
boto3
httpx
gTTS==2.5.1
aiofiles==23.2.1
